Converted some useless Python packages to modules.
