+++ src/wloc/backends/mozilla.py
@@ -4,7 +4,7 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from ..google import BackendGoogle
+from .google import BackendGoogle
 
 
 class BackendMozilla(BackendGoogle):
