+++ src/wloc/exceptions.py
