+++ src/wloc/fetchers/linux/__init__.py
@@ -4,7 +4,7 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from ..native.netman import NetworkManagerNativeAPI
+from .netman import NetworkManagerNativeAPI
 from ...fetchers import FetcherCommon
