+++ src/wloc/fetchers/windows/__init__.py
@@ -4,7 +4,7 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from ..native.wlan import WlanNativeAPI
+from .wlan import WlanNativeAPI
 from ...fetchers import FetcherCommon
