+++ src/wloc/fetchers/windows/structures.py
